Refactored action.rs into multiple files
